+++ src/EntityGraphQL/Compiler/GqlNodes/IGraphQLNode.cs
@@ -6,8 +6,17 @@ namespace EntityGraphQL.Compiler
 {
     public interface IGraphQLNode
     {
+        /// <summary>
+        /// Name of the field
+        /// </summary>
         string Name { get; }
+        /// <summary>
+        /// The expression that represents the field. This will be the context for the next field selection
+        /// </summary>
         Expression? NextFieldContext { get; }
+        /// <summary>
+        /// Parent field. e.g. if we have a field manger like in `people { manager }` then the parent is people
+        /// </summary>
         IGraphQLNode? ParentNode { get; }
         ParameterExpression? RootParameter { get; }