+++ src/EntityGraphQL/Schema/Attributes/GraphQLOneOfAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace EntityGraphQL.Schema
+{
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = true, Inherited = true)]
+    public class GraphQLOneOfAttribute : Attribute
+    {
+    }
+}